requirements.txt
fastapi
pyyaml==6.0.3
mediapipe==0.10.21
opencv-contrib-python==4.11.0.86
uvicorn==0.41.0